# openapi.yml
openapi: 3.0.0
info:
  title: Trivia API
  description: API for a trivia game
  version: 1.0.0
servers:
  - url: http://localhost:8083
    description: Development server
paths:
...
  /games:
    get:
      summary: Get games
      operationId: getGames
      parameters:
        - name: roomId
          in: query
          required: false
          schema:
            type: integer
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Games retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Game"

    post:
      summary: Create a game with the players of a room
      operationId: createGame
      parameters:
        - name: roomId
          in: query
          required: true
          schema:
            type: integer
        - name: rounds
          in: query
          required: true
          schema:
            type: integer
        - name: timePerRound
          in: query
          required: true
          schema:
            type: integer
        - name: questionsPerRound
          in: query
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Game created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
        "400":
          description: Invalid room id or not enough questions
        "401":
          description: Player not authenticated
        "403":
          description: Only the host can create a game
        "409":
          description: Cannot create a new game during a game
  /games/{id}:
    get:
      summary: Get the details of a game
      operationId: getGame
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Game retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
        "404":
          description: Game not found

    delete:
      summary: Delete a game
      operationId: deleteGame
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Game deleted
        "401":
          description: Player not authenticated
        "403":
          description: Only the host can delete the game
        "404":
          description: Game not found
        "409":
          description: Cannot delete an ended game
...
  /questions:
    get:
      summary: Get the questions of a round (correct answers hidden until the round ends)
      operationId: getQuestions
      parameters:
        - name: roundId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Questions retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
        "403":
          description: Round has not started yet
        "404":
          description: Round not found
  /answers:
    post:
      summary: Submit the answer to a question
      operationId: submitAnswer
      parameters:
        - name: roundId
          in: query
          required: true
          schema:
            type: integer
        - name: questionId
          in: query
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                answer:
                  type: string
              required:
                - answer
      responses:
        "200":
          description: Answer submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Answer"
        "401":
          description: Player not authenticated
        "403":
          description: Round has already ended or player is not in the room or in a team
        "404":
          description: Round or question not found

    get:
      summary: >
        Get answers to a question. Without playerId, returns all answers after the round ends.
        With playerId, returns that player's answer (only accessible to players on the same team).
      operationId: getAnswers
      parameters:
        - name: roundId
          in: query
          required: true
          schema:
            type: integer
        - name: questionId
          in: query
          required: true
          schema:
            type: integer
        - name: playerId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Answer(s) retrieved
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: "#/components/schemas/Answer"
                  - $ref: "#/components/schemas/Answer"
        "401":
          description: Player not authenticated (required when playerId is given)
        "403":
          description: Round has not ended yet, or cannot view an answer from a player in another team
        "404":
          description: Round, question, player or answer not found
components:
  schemas:
...
    Game:
      type: object
      properties:
        id:
          type: integer
        roomId:
          type: integer
        createdAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
...
    Question:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        question:
          type: string
        mediaUrl:
          type: string
          nullable: true
        options:
          type: array
          items:
            type: string
        correctAnswers:
          type: array
          items:
            type: string
    Answer:
      type: object
      properties:
        id:
          type: integer
        roundId:
          type: integer
        questionId:
          type: integer
        playerId:
          type: integer
        teamId:
          type: integer
        createdAt:
          type: string
          format: date-time
        answer:
          type: string